+++ test/IfElseTest.php
@@ -2,8 +2,7 @@
 
 namespace PhpFp\Combinators\Test;
 
-use PhpFp\Combinators;
-
+use function PhpFp\if_else;
 use function PhpFp\k;
 
 class IfElseTest extends \PHPUnit_Framework_TestCase
@@ -11,8 +10,15 @@ class IfElseTest extends \PHPUnit_Framework_TestCase
     public function testIfElse()
     {
         $isOdd = function ($x) { return $x % 2 === 1; };
+        $odd = function ($x) { return $x * 3 + 1; };
+        $even = function ($x) { return $x / 2; };
 
-        $f = Combinators::ifElse(
+        $collatz = if_else($isOdd, $odd, $even);
+
+        $this->assertSame(10, $collatz(3));
+        $this->assertSame(1, $collatz(2));
+
+        $f = if_else(
             $isOdd,
             k('Oops!'),
             function ($x) {